back/conquest_back/requirements.txt
asgiref==3.8.1
boto3==1.34.141
botocore==1.34.141
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
DateTime==5.5
defusedxml==0.8.0rc2
diff-match-patch==20230430
Django==5.0.6
django-allauth==0.63.2
django-cors-headers==4.3.1
django-extensions==3.2.3
django-import-export==4.1.0
django-multiselectfield==0.1.12
django-nested-inline==0.4.6
django-rest==0.8.7
django-ses==4.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
google-api-core==2.10.2
google-api-python-client==2.32.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.63.2
gunicorn==22.0.0
httplib2==0.22.0
idna==3.7
jmespath==1.0.1
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2024.1
requests==2.32.2
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.2
sentry-sdk==2.5.1
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tablib==3.5.0
timedelta==2020.12.3
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
zope.interface==6.4.post2
